+++ docs/_pages/releases.md
@@ -12,6 +12,7 @@ sidebar:
 ### What's new
 
 * You can mark all assertions in a class as custom assertions using the `[CustomAssertions]` attribute - [#118](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/118)
+* Added `That[DoNot]HaveAccessModifier` method for filtering types - [#143](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/143)
 
 ### Improvements
 * Provide access to `AssertionChain` that assertion classes were initialized with - [#138](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/138)